tools/pyproject.toml
[project]
name = "lobotomy"
version = "0.1.0"
description = "You don't need to remember, it does. Indexing, search, and management tools for a personal LLM wiki."
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "sentence-transformers>=3.0",
    "sqlite-vec>=0.1.6",
    "numpy>=1.26",
    "tomli>=2.0; python_version < '3.12'",
    "pyyaml>=6.0",
    "mcp[cli]>=1.0",
]

[project.scripts]
lobotomy = "lobotomy.cli:cli"
lobotomy-mcp = "lobotomy.mcp_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lobotomy"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
